Youtube Service
This serves as a wrapper for interacting with the YouTube API to retrieve comments and video information. The class takes the URL of a YouTube video as input and extracts the video ID from it. It then uses the Google API Client Library for Python to build the service and set the API key.